fix: exam query in edit search exam web and mobile
